+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## 4.1.0
 - 新增 element 元素支持，可以自定义列元素属性子元素
 - Grid<D> 数据结构泛型支持
+- Column.sortable 支持 false 参数关闭排序
+- Column.select 支持 false 参数关闭 select
+- Column.dataElement 方法支持设置数据 td 元素
+- Grid.column 方法增加参数可以快捷关闭 select
+- Element 增加 class, style 方法可以快捷设置样式表和样式
+- 分离 Element 为独立项目 element-easy-builder
 
 ## 4.0.3
 - html 渲染使用 page=页 分页